backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.12.0
qdrant-client==1.7.1
supabase==2.3.0
python-docx==1.1.0
werkzeug==3.0.1
langchain-text-splitters>=1.0.0
azure-ai-documentintelligence>=1.0.0
azure-cognitiveservices-speech>=1.32.0
requests>=2.31.0


